extract QualifiedConversationList to file
